create Manager
